generating mini backend bundles for automate HA
